:recycle: move `Store` trait out of API
The `Store` trait is now only used in `drmemd`.
