+++ drivers/drmem-drv-tplink/Cargo.toml
@@ -45,4 +45,4 @@ serde_json.workspace = true
 serde_json.default-features = false
 serde_json.features = ["std"]
 
-drmem-api = { path = "../../drmem-api", version = "0.4" }
+drmem-api = { path = "../../drmem-api", version = "0.5" }
