+++ drivers/drmem-drv-weather-wu/Cargo.toml
@@ -36,4 +36,4 @@ reqwest.default-features = false
 weather-underground.version = "0.1"
 weather-underground.default-features = false
 
-drmem-api = { path = "../../drmem-api", version = "0.4" }
+drmem-api = { path = "../../drmem-api", version = "0.5" }
